+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2024-01-10
 ### Upgraded
 - Keypop Calypso Crypto Legacy SAM API `0.3.0` -> `0.4.0` (management of advanced SAM unlocking)
 
@@ -52,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2022-12-12
 This is the initial release.
 
-[unreleased]: https://github.com/eclipse/keyple-card-calypso-crypto-legacysam-java-lib/compare/0.4.0...HEAD
+[unreleased]: https://github.com/eclipse/keyple-card-calypso-crypto-legacysam-java-lib/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/eclipse/keyple-card-calypso-crypto-legacysam-java-lib/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/eclipse/keyple-card-calypso-crypto-legacysam-java-lib/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/eclipse/keyple-card-calypso-crypto-legacysam-java-lib/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/eclipse/keyple-card-calypso-crypto-legacysam-java-lib/compare/0.1.0...0.2.0